fix : 타입 변경및 이미 테이블에 있는 컬럼 활용

+++ backend/src/main/java/com/quadrant/travelshoot/domains/stay/dto/response/StayTrendingResponse.java
@@ -28,7 +28,7 @@ public class StayTrendingResponse {
     private String checkOutTime;
  
     private String stayType;
-    private Long reviewCount;
+    private Integer reviewCount;
     private Long recent7DaysReservations;
     private Long recent7DaysViews;
